+++ ReleaseNotes.md
@@ -29,6 +29,7 @@
     * Added CollectionViewBlockDelegate for handling UICollectionView item selection and related events.
   * ViewComponents:
     * Added EmbededViewControllerTableViewCell, allowing to embed a UIViewController inside a TableViewCell with ease.
+    * Added CollectionViewColumnLayout, allowing to have a collection view with columns.
   * Views:
     * Added DesignableView for basic views that must be '@IBDesignable'